build.sh run.sh

+++ build.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
-if [ -n "`nvm list|grep 18.`" ]; then
-  nvm current|grep v18 || (nvm use 18.14.2 && sleep 2)
+nodeV=`nvm list|grep -E '\s+18\.'`
+if [ -n "${nodeV}" ]; then
+  nvm current|grep -E '\s+18\.' || (nvm use $nodeV && sleep 2)
 fi
-npm install
+npm install --registry=https://registry.npmmirror.com
 npm run build
+
+if [ $? -eq 1 ]; then
+  exit 1
+fi
 rm -rf dist.tar
 tar cvf dist.tar dist
